Backport the SDL3 refresh-rate printf fix into the vendored runtime
SDL3 made SDL_DisplayMode.refresh_rate a float (SDL_video.h:146) where SDL2
had an int, so passing it to %d is undefined: the panel-rate line printed a
different garbage value every launch (21481008 Hz, 12496848 Hz) and hid the
real reading. With %.2f it reports 30.00 Hz, which is what the host display
was actually set to and what capped the game at 0.50x.

Upstream already fixed this -- it prints %.1f and reworked the pacer entirely
in ed54c668 -- so this is a backport, needed only because UPSTREAM.json pins
psxrecomp 632 commits behind. Do not send it upstream.

Also correct PORTING.md: upstream tracks rabbitizer's 21 generated headers in
git, so their absence is this repo's vendoring pruning them, not an upstream
gap. Record the one-shot refresh-rate read and the settings.toml precedence.

+++ psxrecomp/runtime/src/main.cpp
@@ -12004,11 +12004,11 @@ session_reboot:
             g_host_refresh_hz = host_hz;
             if (host_hz >= 58.8 && host_hz <= 61.2) {
                 g_frame_period_ms = 1000.0 / host_hz;
-                std::printf("psxrecomp: sync-to-host-refresh: pacing to %d Hz panel "
-                            "(%.4f ms/frame)\n", dm.refresh_rate, g_frame_period_ms);
+                std::printf("psxrecomp: sync-to-host-refresh: pacing to %.2f Hz panel "
+                            "(%.4f ms/frame)\n", (double)dm.refresh_rate, g_frame_period_ms);
             } else {
-                std::printf("psxrecomp: host panel %d Hz not ~60 Hz; keeping PSX "
-                            "59.94 Hz pacing\n", dm.refresh_rate);
+                std::printf("psxrecomp: host panel %.2f Hz not ~60 Hz; keeping PSX "
+                            "59.94 Hz pacing\n", (double)dm.refresh_rate);
             }
         }
     }
